ml-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# PDF and document processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# NLP and ML libraries - using pre-compiled wheels for Windows compatibility
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
nltk==3.8.1
sentence-transformers==2.2.2

# Text processing
regex==2023.10.3
textstat==0.7.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data validation
pydantic==2.5.0

# Logging
loguru==0.7.2

# CORS
python-cors==1.7.0
